requirements.txt
streamlit
python-dotenv
groq
chromadb
langchain
langchain-community
langchain-core
langchain-ollama
ollama
pydantic
typing_extensions
tqdm
numpy
